src/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.2
defusedxml==0.8.0rc2
Django==4.2.6
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
django-jazzmin==2.6.0
django-multiselectfield==0.1.12
django-phonenumber-field==7.2.0
django-prometheus==2.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-yasg==1.21.7
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
kombu==5.3.5
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.30
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
twilio==8.13.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4